fix(architecture): isolate ingestion evidence proof assembly
