+++ src/services/event_replay_service/app/proof_builders/__init__.py
@@ -1,0 +1,1 @@
+"""Typed evidence proof assembly for event-replay supportability surfaces."""
